[#1087] Renamed enums file.
