+++ frontend/src/enums/DetailsLevel.ts
